fix: match the accumulator key exactly and split the dashboard tests
A context name may contain a colon, so the prefix check that evicted
stale-generation accumulators for "prod" also took the one for
"prod:staging". Compare the whole key instead.

The concurrent waitForSync test gets a start barrier. Without it the
second caller could arrive after the first had already timed out, which
the load-then-store code it guards against would also pass.

commands_dashboard_test.go was over the 800-line cap before this branch
touched it. The accumulator and fan-out groups move to their own file,
which brings it back under.
